+++ rampy/tests/test_baseline.py
@@ -1,8 +1,5 @@
 import unittest
 
-import sys
-sys.path.append("../../")
-
 import numpy as np
 import scipy